Move PositionMotion class from Moco/Components to Simulation library and fix call sites and bindings
