+++ Bindings/Python/swig/python_simulation.i
@@ -131,6 +131,7 @@ MODEL_ADOPT_HELPER(Controller);
 // the following line:
 %template(_StdVectorState) std::vector<SimTK::State>;
 
+moco_unique_ptr(OpenSim::PositionMotion);
 
 // Include all the OpenSim code.
 // =============================
