+++ OpenSim/Moco/RegisterTypes_osimMoco.cpp
@@ -19,7 +19,6 @@
 
 #include "Components/AccelerationMotion.h"
 #include "Components/DiscreteForces.h"
-#include "Components/PositionMotion.h"
 #include "Components/StationPlaneContactForce.h"
 #include "MocoBounds.h"
 #include "MocoCasADiSolver/MocoCasADiSolver.h"
@@ -109,7 +108,6 @@ OSIMMOCO_API void RegisterTypes_osimMoco() {
         Object::registerType(AckermannVanDenBogert2010Force());
         Object::registerType(MeyerFregly2016Force());
         Object::registerType(EspositoMiller2018Force());
-        Object::registerType(PositionMotion());
 
         Object::registerType(DiscreteForces());
         Object::registerType(AccelerationMotion());
