+++ OpenSim/Simulation/osimSimulation.h
@@ -125,6 +125,7 @@
 #include "StatesTrajectory.h"
 #include "StatesTrajectoryReporter.h"
 #include "TableProcessor.h"
+#include "PositionMotion.h"
 #include "OpenSense/OpenSenseUtilities.h"
 #include "OpenSense/IMU.h"
 #include "SimulationUtilities.h"
